New versions of deleteTask methods in Controller & Models.

+++ app/models/ToDoModel_mongo.php
@@ -113,14 +113,20 @@ class ToDoModel_mongo {
     }
 
     // DELETE: method that deletes a task from Mongo DataBase
-    public function deleteTask(int $id): bool { 
+    public function deleteTask(int $id): bool {
+
+        $task = $this -> getTaskById($id);
+
+        if(!$task) {
+            return false;  
+        } 
 
         $result =  $this->_collection->deleteOne(['id' => $id]); 
 
         if (!$result->getDeletedCount()) { 
             return false;   
         }
-        //repassar?
+        
         return true;    
             
     }
